+++ notification-service/src/main/java/com/aditya/ecommerce/notification/strategy/NotificationChannel.java
@@ -1,0 +1,7 @@
+package com.aditya.ecommerce.notification.strategy;
+
+public enum NotificationChannel {
+    EMAIL,
+    SMS,
+    PUSH
+}