+++ notification-service/src/main/java/com/aditya/ecommerce/notification/strategy/NotificationStrategy.java
@@ -1,0 +1,10 @@
+package com.aditya.ecommerce.notification.strategy;
+
+import com.aditya.ecommerce.notification.event.OrderCreatedEvent;
+
+public interface NotificationStrategy {
+
+    NotificationChannel channel();
+
+    void send(OrderCreatedEvent event);
+}