+++ order-service/src/main/java/com/aditya/ecommerce/order/domain/Order.java
@@ -27,6 +27,8 @@ public class Order {
     @Column(nullable = false)
     private String customerUsername;
 
+    private String customerEmail;
+
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private OrderStatus status;